修复部分BUG...

+++ src/main/java/net/lightwing/qyshm_web/controller/admin/QBannerController.java
@@ -36,7 +36,7 @@ public class QBannerController {
 
     @RequestMapping("insert")
     @ResponseBody
-    public Wrapper insert(QBanner qBanner, @RequestParam("imgFile") MultipartFile file) {
+    public Wrapper insert(QBanner qBanner, @RequestParam(value = "imgFile",required = false) MultipartFile file) {
         if (StringUtils.isNotBlank(file.getOriginalFilename())) {
             qBanner.setImgpath(null);
             if (qBanner.getImgpath() != null) {
@@ -54,20 +54,21 @@ public class QBannerController {
 
     @RequestMapping("update")
     @ResponseBody
-    public Wrapper update(QBanner qBanner, @RequestParam("imgFile") MultipartFile file) {
-
-        if (StringUtils.isNotBlank(file.getOriginalFilename())) {
-            qBanner.setImgpath(null);
-            if (qBanner.getImgpath() != null) {
-                String deletepath = qBanner.getImgpath().substring(1, qBanner.getImgpath().length());
-                UPLOAD.deleteFile(deletepath);
-            }
-            Map<String, Object> upload = UPLOAD.UPLOADFILE(file);
-            if ((int) upload.get("code") == 200) {
-                qBanner.setImgpath("/pictures/" + upload.get("filename"));
-                QBanner result = qBannerService.selectById(qBanner.getBid());
-                if (result.getImgpath() != null) {
-                    UPLOAD.deleteFile(result.getImgpath());
+    public Wrapper update(QBanner qBanner, @RequestParam(value = "imgFile",required = false) MultipartFile file) {
+        if(file!=null) {
+            if (StringUtils.isNotBlank(file.getOriginalFilename())) {
+                qBanner.setImgpath(null);
+                if (qBanner.getImgpath() != null) {
+                    String deletepath = qBanner.getImgpath().substring(1, qBanner.getImgpath().length());
+                    UPLOAD.deleteFile(deletepath);
+                }
+                Map<String, Object> upload = UPLOAD.UPLOADFILE(file);
+                if ((int) upload.get("code") == 200) {
+                    qBanner.setImgpath("/pictures/" + upload.get("filename"));
+                    QBanner result = qBannerService.selectById(qBanner.getBid());
+                    if (result.getImgpath() != null) {
+                        UPLOAD.deleteFile(result.getImgpath());
+                    }
                 }
             }
         }
@@ -79,7 +80,7 @@ public class QBannerController {
     @ResponseBody
     public Wrapper delete(Integer bid) {
         QBanner qBanner = qBannerService.selectById(bid);
-        if (qBanner.getImgpath() != null) {
+        if (qBanner.getImgpath()!=null) {
             UPLOAD.deleteFile(qBanner.getImgpath());
         }
         qBannerService.delete(bid);
